+++ cogs/bot_info.py
@@ -1,23 +1,21 @@
-import os
 from datetime import datetime
 
-import discord
-from discord.ext.commands import Cog, command
+from discord import Embed
+from discord.ext.commands import Cog, slash_command
 
 
 class BotInfo(Cog):
     def __init__(self, bot):
         self.bot = bot
         self.start_time = datetime.now()
-        self.prefix = '!'
-        if 'BOT_PREFIX' in os.environ:
-            self.prefix = os.environ['BOT_PREFIX']
+        self.prefix = '/'
 
-    @command()
+    # bot info
+    @slash_command(description="Displays info and more commands")
     async def info(self, ctx):
         """Sends a help message"""
         info = await self.bot.application_info()
-        embed = discord.Embed(
+        embed = Embed(
             title='Plant ID Bot Info',
             description=f'{info.description}',
             colour=0x41c03f
@@ -26,8 +24,8 @@ class BotInfo(Cog):
             value='Shows this message',
             inline=True
         ).add_field(
-            name=f'`{self.prefix}idhelp`',
-            value='Gives more detailed help on the plant ID bot',
+            name=f'`{self.prefix}help`',
+            value='Gives detailed help on the plant ID bot',
             inline=True 
         ).add_field(
             name=f'`{self.prefix}stats`',
@@ -42,12 +40,13 @@ class BotInfo(Cog):
             value='Links to the bot\'s code repo',
             inline=True
         ).set_footer(text=f'Made by {info.owner}')
-        await ctx.send(embed=embed)
+        await ctx.respond(embed=embed)
 
-    @command()
+    # bot stats
+    @slash_command(description="Displays bot stats")
     async def stats(self, ctx):
         info = await self.bot.application_info()
-        embed = discord.Embed(
+        embed = Embed(
             title=f'{info.name}',
             description="Stats etc.",
             colour=0x1aaae5,
@@ -68,37 +67,28 @@ class BotInfo(Cog):
             value=f'{round(self.bot.latency * 1000, 2)}ms',
             inline=True
         ).set_footer(text=f'Made by {info.owner}')
-        await ctx.send(embed=embed)
+        await ctx.respond(embed=embed)
 
     # Sends the link to the bot's GitHub repo
-    @command()
+    @slash_command(description="A link to the bot's source code repository")
     async def source(self, ctx):
         url = 'https://github.com/TheRealOwenRees/plantID_discordbot'
-        embed = discord.Embed(
-            title=f'View the repo and report issues at:',
-            description=f'{url}',
-            colour=0x1aaae5
-        )
-        await ctx.send(embed=embed)        
+        await ctx.respond(f'View source code and report issues at:\n\n{url}')
 
     # Sends a bot invite link
-    @command()
+    @slash_command(description="Display a bot invite link")
     async def invite(self, ctx):
         url = 'https://discord.com/api/oauth2/authorize?client_id=948227126094598204&permissions=19520&scope=bot'
-        embed = discord.Embed(
-            title=f'Invite the bot to your server:',
-            description=f'[Click here]({url})',
-            colour=0x1aaae5
-        )
-        await ctx.send(embed=embed)
+        await ctx.respond(f'Invite the bot to your server:\n\n[Click here]({url})')
 
     # list servers that the bot has been invited to
-    @command()
+    @slash_command(description="All servers that this bot belongs to")
     async def servers(self, ctx):
         servers = list(self.bot.guilds)
-        embed = discord.Embed(
-            title=f'Connected on {str(len(servers))} servers:',
+        server_wording = 'servers' if len(servers) > 1 else 'server'
+        embed = Embed(
+            title=f'Connected on {len(servers)} {server_wording}:',
             description='\n'.join(guild.name for guild in self.bot.guilds),
             colour=0x1aaae5
         )
-        await ctx.send(embed=embed)
+        await ctx.respond(embed=embed)
